+++ packages/harness/src/core/mcp/auth/index.ts
@@ -1,5 +1,4 @@
 export {
-  DEFAULT_MCP_ACCESS_TOKEN_TTL_SECONDS,
   buildMcpAuthorizationUrl,
   completeMcpAuthorization,
   createMcpOAuthClient,
@@ -14,6 +13,7 @@ export type {
   ResolveMcpAuthResult,
 } from './mcpDcr';
 export {
+  DEFAULT_MCP_ACCESS_TOKEN_TTL_SECONDS,
   MCP_OAUTH_CALLBACK_PATH,
   mcpAuthorizationServerMetadata,
   mcpAuthorizationServerOrigin,
